Refactor(OKRDisplay): Moved to Display component (front, action -> state, db)
